objeto do serialW adicionado; objeto do imageView adicionado
